+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/BusinessViewModel.java
@@ -12,6 +12,7 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 import com.sdzshn3.android.news247.Fragments.BusinessNewsFragment;
+import com.sdzshn3.android.news247.Repositories.BusinessRepository;
 import com.sdzshn3.android.news247.Retrofit.ApiService;
 import com.sdzshn3.android.news247.Retrofit.Article;
 import com.sdzshn3.android.news247.Retrofit.Client;
@@ -20,54 +21,19 @@ import com.sdzshn3.android.news247.Retrofit.NewsModel;
 import java.util.List;
 
 public class BusinessViewModel extends AndroidViewModel {
-    private static final MutableLiveData<List<Article>> data = new MutableLiveData<>();
 
-    private ApiService apiService;
-    private Call<NewsModel> call;
+    private BusinessRepository repository;
 
     public BusinessViewModel(@NonNull Application application) {
         super(application);
-        apiService = Client.getApiService();
-        loadData();
+        repository = new BusinessRepository();
     }
 
     public LiveData<List<Article>> getData() {
-        return data;
-    }
-
-    private void loadData() {
-        call = apiService.getResponse(BusinessNewsFragment.URL);
-        call.enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("BusinessViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        return repository.getData();
     }
 
     public void refresh() {
-        call = apiService.getResponse(BusinessNewsFragment.URL);
-        call.clone().enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("BusinessViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository.refresh();
     }
 }
